Refactor cognition layer into new mind package
