+++ Server/app/tests/test_conversation_service.py
@@ -14,15 +14,11 @@ SERVER_ROOT = Path(__file__).resolve().parents[2]
 if str(SERVER_ROOT) not in sys.path:
     sys.path.insert(0, str(SERVER_ROOT))
 
-core_stub = types.ModuleType("core")
-core_stub.__path__ = []  # type: ignore[attr-defined]
-sys.modules.setdefault("core", core_stub)
+mind_stub = types.ModuleType("mind")
+mind_stub.__path__ = []  # type: ignore[attr-defined]
+sys.modules.setdefault("mind", mind_stub)
 
-llm_stub = types.ModuleType("core.llm")
-llm_stub.__path__ = []  # type: ignore[attr-defined]
-sys.modules.setdefault("core.llm", llm_stub)
-
-llama_stub = types.ModuleType("core.llm.llama_server_process")
+llama_stub = types.ModuleType("mind.llama_server_process")
 
 
 class _StubLlamaServerProcess:  # pragma: no cover - test scaffolding
@@ -30,9 +26,9 @@ class _StubLlamaServerProcess:  # pragma: no cover - test scaffolding
 
 
 llama_stub.LlamaServerProcess = _StubLlamaServerProcess
-sys.modules.setdefault("core.llm.llama_server_process", llama_stub)
+sys.modules.setdefault("mind.llama_server_process", llama_stub)
 
-llm_client_stub = types.ModuleType("core.llm.llm_client")
+llm_client_stub = types.ModuleType("mind.llm_client")
 
 
 class _StubLlamaClient:  # pragma: no cover - test scaffolding
@@ -44,15 +40,14 @@ class _StubLlamaClient:  # pragma: no cover - test scaffolding
 
 
 llm_client_stub.LlamaClient = _StubLlamaClient
-sys.modules.setdefault("core.llm.llm_client", llm_client_stub)
+sys.modules.setdefault("mind.llm_client", llm_client_stub)
 
 from app.services.conversation_service import ConversationService
 
 for name in [
-    "core.llm.llama_server_process",
-    "core.llm.llm_client",
-    "core.llm",
-    "core",
+    "mind.llama_server_process",
+    "mind.llm_client",
+    "mind",
 ]:
     sys.modules.pop(name, None)
 
@@ -64,6 +59,8 @@ class ProcessMock:
         self.wait_ready = mock.Mock(return_value=True)
         self.terminate = mock.Mock(side_effect=self._terminate)
         self.is_running = mock.Mock(side_effect=self._is_running)
+        self.poll = mock.Mock(return_value=None)
+        self.poll_health = mock.Mock(side_effect=self._poll_health)
 
     def _start(self) -> None:
         self._running = True
@@ -73,6 +70,10 @@ class ProcessMock:
 
     def _is_running(self) -> bool:
         return self._running
+
+    def _poll_health(self, *_args: Any, **_kwargs: Any) -> bool:
+        self.terminate()
+        return True
 
 
 def _build_service(
