+++ Server/core/README.md
@@ -1,0 +1,4 @@
+# Core Layer
+
+Core provides physical interfaces and low-level control only. Cognitive
+functions have been moved to `mind/`.
