+++ Server/core/llm/__init__.py
@@ -1,0 +1,1 @@
+from mind import *  # legacy redirect
