+++ Server/mind/README_Qwen(Windowns).md
