+++ Server/mind/README_Qwen.md
