+++ Server/tests/test_app_runtime_conversation_integration.py
@@ -22,9 +22,11 @@ vosk_stub = types.ModuleType("vosk")
 core_stub = types.ModuleType("core")
 core_stub.__path__ = [str(SERVER_ROOT / "core")]
 sys.modules["core"] = core_stub
-importlib.import_module("core.llm.llama_server_process")
-sys.modules.setdefault("core.llm", sys.modules.get("core.llm"))
-setattr(sys.modules["core"], "llm", sys.modules["core.llm"])
+
+mind_stub = types.ModuleType("mind")
+mind_stub.__path__ = [str(SERVER_ROOT / "mind")]
+sys.modules["mind"] = mind_stub
+importlib.import_module("mind.llama_server_process")
 
 requests_stub = types.ModuleType("requests")
 
@@ -48,9 +50,9 @@ requests_stub.Response = _StubRequestsResponse
 sys.modules.setdefault("requests", requests_stub)
 
 def teardown_module() -> None:
-    sys.modules.pop("core.llm.llama_server_process", None)
-    sys.modules.pop("core.llm", None)
-    sys.modules.pop("core.VoiceInterface", None)
+    sys.modules.pop("mind.llama_server_process", None)
+    sys.modules.pop("mind.interface.voice_interface", None)
+    sys.modules.pop("mind", None)
     sys.modules.pop("core", None)
 
 def _no_op(*_args, **_kwargs) -> None:
@@ -62,6 +64,7 @@ sys.modules.setdefault("cv2", cv2_stub)
 sys.modules.setdefault("numpy", numpy_stub)
 sys.modules.setdefault("numpy.typing", numpy_typing_stub)
 sys.modules.setdefault("LedController", led_controller_stub)
+sys.modules.setdefault("core.LedController", led_controller_stub)
 sys.modules.setdefault("sounddevice", sounddevice_stub)
 sys.modules.setdefault("vosk", vosk_stub)
 
@@ -266,7 +269,7 @@ def _build_runtime_with_conversation() -> tuple[
     llama_process = StubLlamaProcess()
     led_handler = StubLedHandler()
 
-    from core.VoiceInterface import ConversationManager as _ConversationManager
+    from mind.interface.voice_interface import ConversationManager as _ConversationManager
 
     stop_event_ref: Dict[str, threading.Event] = {}
