+++ Server/tests/test_llama_server_process.py
@@ -1,6 +1,5 @@
 import sys
 import time
-import types
 from pathlib import Path
 from unittest import mock
 
@@ -10,12 +9,8 @@ SERVER_ROOT = Path(__file__).resolve().parents[1]
 if str(SERVER_ROOT) not in sys.path:
     sys.path.insert(0, str(SERVER_ROOT))
 
-core_stub = types.ModuleType("core")
-core_stub.__path__ = [str(SERVER_ROOT / "core")]
-sys.modules.setdefault("core", core_stub)
-
-from core.llm.llama_server_process import LlamaServerProcess
-import core.llm.llama_server_process as llama_process_module
+from mind.llama_server_process import LlamaServerProcess
+import mind.llama_server_process as llama_process_module
 
 
 def test_start_is_non_blocking(dummy_binary: Path, dummy_model: Path) -> None:
@@ -86,7 +81,7 @@ def test_poll_health_logs_backoff(
     caplog.set_level("INFO")
 
     monkeypatch.setattr(
-        "core.llm.llama_server_process.urllib_request.urlopen",
+        "mind.llama_server_process.urllib_request.urlopen",
         mock.Mock(side_effect=llama_process_module.urllib_error.URLError("boom")),
     )
